+++ src/test/cli_test.ts
@@ -14,19 +14,49 @@
 
 import {assert} from 'chai';
 import {run as cliRun} from '../cli';
+import intercept = require('intercept-stdout');
 
 suite('cli', () => {
-
-  test('unknown cmd parameter should not throw exception', () => {
+  /**
+   * Runs the CLI with the given args.
+   *
+   * Why not do this setup and teardown in the standard  setup() and teardown()
+   * methods? Because you don't want to trap stdout at test end, as that would
+   * trap mocha's output.
+   *
+   * TODO(rictic): look into using spawn() to run polyserve in another process
+   *     instead. That way we could actually get it started and run it for a
+   *     while, hit it with requests, etc.
+   *
+   * @return A Promise of all STDOUT contents that're written during the
+   *     run() function in cli.ts.
+   */
+  async function runCli(args: string[]) {
     const originalArgv = process.argv;
-    process.argv = ['node', 'polyserve', '--unknown-parameter'];
+    process.argv = ['node', 'polyserve'].concat(args);
+    let stdout = '';
+    let unintercept = intercept((txt) => {
+      stdout += txt;
+      return '';
+    });
 
     try {
-      return cliRun();
+      await cliRun();
     } finally {
-      // restore process arguments for other readers
+      unintercept();
       process.argv = originalArgv;
     }
+    return stdout;
+  }
+
+  test('unknown cmd parameter should not throw exception', async() => {
+    const stdout = await runCli(['--unknown-parameter']);
+
+    // Assert that we printed something that looks kinda like help text to
+    // stdout.
+    assert.match(stdout, /A development server for Polymer projects/);
+    assert.match(stdout, /--version/);
+    assert.match(stdout, /--package-name/);
   });
 
 });