+++ src/transform-middleware.ts
@@ -1,17 +1,22 @@
 /**
  * @license
  * Copyright (c) 2016 The Polymer Project Authors. All rights reserved.
- * This code may only be used under the BSD style license found at http://polymer.github.io/LICENSE.txt
- * The complete set of authors may be found at http://polymer.github.io/AUTHORS.txt
- * The complete set of contributors may be found at http://polymer.github.io/CONTRIBUTORS.txt
+ * This code may only be used under the BSD style license found at
+ * http://polymer.github.io/LICENSE.txt
+ * The complete set of authors may be found at
+ * http://polymer.github.io/AUTHORS.txt
+ * The complete set of contributors may be found at
+ * http://polymer.github.io/CONTRIBUTORS.txt
  * Code distributed by Google as part of the polymer project is also
- * subject to an additional IP rights grant found at http://polymer.github.io/PATENTS.txt
+ * subject to an additional IP rights grant found at
+ * http://polymer.github.io/PATENTS.txt
  */
 
-import {RequestHandler, Response, Request} from 'express';
+import {Request, RequestHandler, Response} from 'express';
 import {Transform} from 'stream';
 
-export function transformResponse(transformer: ResponseTransformer): RequestHandler {
+export function transformResponse(transformer: ResponseTransformer):
+    RequestHandler {
   return (req: Request, res: Response, next: () => void) => {
     let ended = false;
 
@@ -27,16 +32,19 @@ export function transformResponse(transformer: ResponseTransformer): RequestHand
     }
 
     const _write = res.write;
-    res.write = function(chunk: Buffer|string, cbOrEncoding?: Function|string, cbOrFd?: Function|string): boolean {
+    res.write = function(
+        chunk: Buffer|string,
+        cbOrEncoding?: Function|string,
+        cbOrFd?: Function|string): boolean {
       if (ended) {
         _write.call(this, chunk, cbOrEncoding, cbOrFd);
         return false;
       }
 
       if (shouldTransform()) {
-        const buffer = (typeof chunk === 'string')
-          ? new Buffer(chunk, cbOrEncoding as string)
-          : chunk;
+        const buffer = (typeof chunk === 'string') ?
+            new Buffer(chunk, cbOrEncoding as string) :
+            chunk;
         chunks.push(buffer);
         return true;
       } else {
@@ -45,8 +53,12 @@ export function transformResponse(transformer: ResponseTransformer): RequestHand
     };
 
     const _end = res.end;
-    res.end = function(chunk?: Buffer|string, cbOrEncoding?: Function|string, cbOrFd?: Function|string): boolean {
-      if (ended) return false;
+    res.end = function(
+        chunk?: Buffer|string,
+        cbOrEncoding?: Function|string,
+        cbOrFd?: Function|string): boolean {
+      if (ended)
+        return false;
       ended = true;
 
       if (shouldTransform()) {
@@ -70,7 +82,6 @@ export function transformResponse(transformer: ResponseTransformer): RequestHand
 }
 
 export interface ResponseTransformer {
-
   /**
    * Returns `true` if this transformer should be invoked.
    * Transformers should only look at headers, do not call res.write().